PYIC-3599 Moved vchelper to helpers package.
